Able to Register User and Login and Edit the App.vue and LandingView.vue
For App.vue, now is able to detect whether user has logged in and show navbar based on logged in situation. Logout button has been added to navbar and onclick will signout the user, hide the navbar and return to landingview. Style has been edited too.
For Login and Registration, now is working and user will be redirect to respective home page.